requirements.txt
pyyaml==5.3.1
paho-mqtt==1.6.1